Rename ls to objects
