+++ objects/objects.go
@@ -1,4 +1,4 @@
-package ls
+package objects
 
 import (
 	"io/ioutil"
